Add projected updates on the client
Given a state snapshot from the server, the client is now able to
calculate the game state at a given frame in the future. This avoids
having to stream a state update for every frame.

Note: There are some issues with the strategy reguarding floating point
determisism. It might be neccsary to switch to interger types for ship
counts. This does complicate rounding though.