Fix client crashes caused by outdated buckets
Modification buckets hold ship count deltas that are caused by past
moves, that are stored until the appropriate time to apply them. They
are created when the move is created. They are applied during the
process of stepping to a target time grater than the move.

The issue arrises when the state gets moved passed the modification
bucket application time without processing the modificatio bucket. This
can happen when state snapshots are recieved from the server.

Example:
If there is a move a t=10, that will result in -5 ships at t=50, a
modification bucket will queued for application at t=50. Lets say the
last requested time on the client is now t=49, and a state for t=51 has
been loaded from the server. The next requested time will be t=51.
The modification bucket for t=50 is still in the queue, however it has
already been applied by the server, and cannot be applied to the current
state again. Therefore we need to pop all buckets less than the current
calculated state time, before applying the buckets.

This bug removes these outdated modification buckets from the front of
the queue before processing modificiation buckets to reach the target
time.