Move ship shader creation out of render loop
Creating and verifying a shader is a slow process that reduces the
framerate. The shader only needs to be initalized once and can be reused
between frames.

I moved the creation out of the render loop into the Game Render
constructor, and then store the output program inside the Game Render.

The speed up was very substantial, doubling the framerate in some tests.