Refactor game move rendering for performance
Previously, the entire game move render setup and tairdown would happen
every frame instead of once per move.

This adds a MoveRender struct that provides a create, render, and drop
methods that allow the render setup to be created, used and destroyed.

The MoveRender structs are stored in the GameRender object while the
move is in play. They are cleaned up once the move is over.