Fix broken match in client handler for GameEvents
The enum variants for GameEvent were not in scope, causing the variant
names to be treated as variables rather than patterns. As a result, the
first branch, which was "Start" would respond to all the events, which
I somehow failed to noticed up until now, despite clippy warning that
the other branches were dead code *facepalm*.

The code has now been updated to a) use the enum prefix to disambiguate
the enum variants, and b) exhaustively match all variants.