Add event handler unsubscribe to game window
When the user exits the game window (to go to another game), the event
handlers remaing attactched to the gameConnectionSingleton, which can
cause crashes (because it will try to update non-existant UI.

`gameConnectionSingleton.onEvent` was updated to return an unsubscribe
function, that can be called to remove the handler.

`GameWindow` was then updated to use these unsubscribe functions to
remove the handlers when the component is destroyed.