Rename Move.time to Move.start_time
start_time is slightly more clear. Moves take a while to be fully
applied since troops need to travel between planets. This makes it more
clear that this is the time of the initating event, as opposed to the
move's end time, or any other time relating to the move.