Fix sync bug for late-joining clients
Joining a game after some moves had been made resulted in the client
showing higher troop counts than the server and the other clients.

The root cause of this bug was line 412 of core/src/game/mod.rs which
set the `galaxy.time` to `game_move.start_time`. Computation is
performed in jumps between moves, rather than on each tick, because most
of the game calculations that would be performed on each tick are
essentially redundant. As a result, the galaxy time is always set to
the the start time of a move when it is jumped to. This works fine for
clients that join at the start of the game and the server, which always
process moves from the present and the future.

Late joining clients, however, process moves from the past. This is
neccsary to correctly compute the planet deltas for troops that are in
transit. When a troop is deployed, it's delta is added to a
"modification bucket" that is applied at the appropriate time. This is
better than calculating the troops movement one frame at a time.

The impact of processing past moves is that the galaxy's time is moved
into the past (since the game move is from the past). Moving the galaxy
into the past causes troop growth to be re-counted over those peroids.

The bug was fixed by only updating the galaxy's time if the game_move's
time was larger. For past moves, we are only interested in future
mod-buckets, since past mod-buckets are already baked into the game
state.