Add server client clock syncing protocol
Creates a protocl for the server and the client to sync their clocks.
The client requests the servers current clock time. When the client
recieves the server's clock time, it will find the difference between
the it's current time and the server's clock time. That difference
represents the delta between the server and client clock time including
latency. This can be used to convert recieved server clock times back to
their original real time.