Refactor game client
I refactored the game client to group state related to the *current*
game as opposed to the game connection in general.

What happened was there was a growing number of fields with option types
named `current_` something. All of these fields are needed at various
points when joining and playing a game. Additionally, when leaving a
game all of these fields needed to be cleaned up. By grouping them, I
could clean them up in one foul swoop additionally, I can enforce that
all the fields are set together, removing the possibility of weird
partial states. I did this by putting the fields into an enum called
ActiveGame that had the variants JoinedGame and Waiting, which contained
structs that represent the state needed while a player is in a game, and
 the state needed to join a game.

Overall, this refactoring wasn't terriby successful, although I decided
to keep it since I think this a step towards a more sucessful
application design. The biggest drawback was that the borrow checker
punishes grouping. Having the fields as indepenant options allowed them
to be borrowed seperately. Grouping them caused them to be borrowed
together, which caused a lot of problems with mutable references. The
borrow checker only permits a single mutable reference, which caused
problems when I tried to borrow one part of one thing and another part
of another thing. I had to add a lot of hacks to work around this, and
the code became really messy and unclear.

I think this problem can be solved though, with even *more* refactoring
and sub-grouping. Specifically, if I move methods on to GameClient's
fields, I can avoid a lot of the excessive borrowing that is the root
of the problem.

Another flaw in the design was moving the state held in Waiting into
JoinedGame. You can't just move data in and out of an enum very easily,
because there will be a moment when nothing is in that enum, which is a
problem if there is a panic. I had to work around it by using mem-repace
and the None variant of the ActiveGame enum.

The mem-replace pattern is covered in detail here:

http://smallcultfollowing.com/babysteps/blog/2018/11/10/after-nll-moving-from-borrowed-data-and-the-sentinel-pattern/