Refactor accessing games list into hook
There is a bit of boilerplate code involved in getting the game list
out of the game client. An event handler has to be added to pickup list
change events so it can set the react state, and that handler needs to
be cleaned up when the component unmounts.

I refactored this boilerplate out into a seperate hook, which reduces
the amount of noise in the components. It also makes the logic
reuseable.