Fix React warning on controlled select element
I was setting selected on the default option element, which was
triggering a warning from React saying I should use the value property
on the parent element.

Using value is preferred because it avoids needing to specify selected
on  all of the child options, which makes setting up a controlled
component easier.