Add join link routing logic
I added code that creates links which users can use to automatically
join games, and the routing logic to support those links. This logic
runs when pages load, and whenever the hash state in changed, checking
for a join link for an open game. If it finds a open game, it sets that
game as the current game.

A potential improvement is flashing an error message if the user users a
link that does not point to a valid game, so they aren't confused if the
join link doesn't work.