Add connection error message and retry
I added logic that displays an error message when the client is unable
to establish a connection to the websocket.  This is useful for helping
users understand why the UI doesn't work.

I also added retry with exponential backoff, so if the connection is
temporarily interrupted the user does not have to refresh the page.

I also refactored the GameConnection object a little bit to move
repeated event triggering code into a method.