Add eslint config and fix discovered errors
I'm am not very big on linters, especially persnickety ones. I have
better things to think about than whether I should use double or single
quote, or have a space between the // and the comment text.

However, eslint does provide some critical linting. Specifically, it can
catch serious react hooks bugs when using the react hooks eslint plugin.
The bug is that if dependency is left out of the dependency list on a
react hook, that state can be stale in the hook, which can cause bugs.

In this case the linter found one such oversight in the code, which
validates the need for it.