Remove mutex around connection Sink
The server was having issues where it would freeze up, most likely due
to Mutex related deadlocking. I did not peruse the root cause of
the issue, instead opting to restructure the program a little. The key
change was to swap out the WebSocket sink, which is not Sync with a
mpsc sender, which is Sync. This removes a large number of locks,
reducing the likelyhood of a deadlock.

I did have to spin up an extra thread per connection to forward the
messages from the mpsc to the WebSocket Sink. This seems like it might
not be neccsary. There may be a way to expose the Sink directly in a
Sync manner. The forwarding approach was choosen for implementation
simplicty. I will revisit this at a later point.