Add selection ring around planets
A friend gave me feedback that the selection mechanics were confusing.
In particular, the fact that you can click on a planet (without
dragging) to select it is not obvious. To make the mechanics more clear,
I've added a selection outline which highlights the currently selected
planet.

While implementing and debugging this feature, I also made some code
quality improvements to the rendering logic. In particular, the code
that renders the ships was setting vertex array object attribtues inside
of the render loop, rather than once during intialization. This is
uneccsary, and has the potential to introduce bugs. I moved this code
into the MoveRender intialization logic, where it belongs. I also added
code to make the move's VAO the active VAO when during the render loop,
which previously wasn't implemented.

I've also created a new abstraction that wraps buffers, and
automatically drops them. I haven't refeactored the existing code to use
the abstraction yet, but I will do so in a future commit.

The selection ring has futher improvements that it needs:
 * Anti-aliasing. There are aliasing artifacts around the edge of the
 ring.
 * Color. I want to make the selection ring match the player color

I think there also needs to be further enhancements to the visual style
 - as it stands right now, it looks incomplete. I'm not entirely sure
what improvements yet.