Hide UI when connection is not open
None of the game UI works when there is no connection. To avoid
presenting the user with a broken interface, I now hide the UI when the
connection is not open. I've also added a 'connecting' dialog that shows
if the connection is taking a long time to establish. This prevents the
user from seeing a blank screen if the connection is slow.